+++ idea_simulator/src/core/block.c
@@ -102,7 +102,7 @@ struct line *add_line_block(struct block *block, struct line *line, int (*replac
 }
 
 
-void update_FIFO(struct block *block, int nb_ways, int entry) {
+void update_FIFO(struct block *block, int nb_ways, long entry) {
   (void) entry;
   int i;
   struct line *line;
@@ -114,7 +114,7 @@ void update_FIFO(struct block *block, int nb_ways, int entry) {
   }
 }
 
-void update_LFU(struct block *block, int nb_ways, int entry) {
+void update_LFU(struct block *block, int nb_ways, long entry) {
   int i;
   struct line *line;
   for (i=0; i<nb_ways; i++) {
@@ -125,7 +125,7 @@ void update_LFU(struct block *block, int nb_ways, int entry) {
   }
 }  
 
-void update_LRU(struct block *block, int nb_ways, int entry) {
+void update_LRU(struct block *block, int nb_ways, long entry) {
   (void) entry;
   int i;
   int max = 0;
